Allow decimal temperatures to be entered

+++ SeaPublicWebsite/DataModels/UserDataModel.cs
@@ -39,7 +39,7 @@ namespace SeaPublicWebsite.DataModels
         
         public int? NumberOfOccupants { get; set; }
         public HeatingPattern? HeatingPattern { get; set; }
-        public int? Temperature { get; set; }
+        public decimal? Temperature { get; set; }
         
         public HasEmailAddress? HasEmailAddress { get; set; }
         public string EmailAddress { get; set; }
